backend-api-service/requirements.txt
python-dotenv
pydantic
pydantic[email]
motor
pymongo
httpx
python-jose
passlib[bcrypt]
bcrypt
dnspython